Extract wlcs::ActiveListeners for reuse
